Update simplescheduler.sh
The code line to remove the file /var/www/html/index.html will cause the container to fail if that file does not exist. 
This happens when running in docker container mode, if you stop/start the container after the first launch.
I suggest adding a conditional statement, simply to check whether the file exists before trying to delete it.